+++ python/tests/team_model_select_test.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Google Inc.
+# Copyright 2017-2018 Google Inc.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -27,20 +27,23 @@ import deepmind_lab
 class TeamModelSelectTest(unittest.TestCase):
 
   def test_with_six_bots(self):
+    player_count = 7
     env = deepmind_lab.Lab(
         'tests/team_model_select', [],
-        config={'botCount': '6'})
+        config={'botCount': str(player_count - 1)})
     env.reset()
     action_spec = env.action_spec()
     action = np.zeros([len(action_spec)], dtype=np.intc)
     event_names = sorted([name for name, _ in env.events()])
-    self.assertEqual(len(event_names), 7)
-    for i, name in enumerate(event_names):
-      self.assertEqual(name, 'skinModified' + str(i))
+    self.assertEqual(len(event_names), player_count * 2)
+    for i in xrange(player_count):
+      self.assertEqual(event_names[i], 'newClientInfo' + str(i + 1))
+      self.assertEqual(event_names[i + player_count],
+                       'skinModified' + str(i + 1))
     env.step(action, 1)
     self.assertEqual(len(env.events()), 0)
     env.reset()
-    self.assertEqual(len(env.events()), 7)
+    self.assertEqual(len(env.events()), player_count * 2)
 
   def test_without_bot(self):
     env = deepmind_lab.Lab(
@@ -50,12 +53,14 @@ class TeamModelSelectTest(unittest.TestCase):
     action_spec = env.action_spec()
     action = np.zeros([len(action_spec)], dtype=np.intc)
     event_names = sorted([name for name, _ in env.events()])
-    self.assertEqual(len(event_names), 1)
-    self.assertEqual(event_names[0], 'skinModified0')
+    self.assertEqual(len(event_names), 2)
+    self.assertEqual(event_names[0], 'newClientInfo1')
+    self.assertEqual(event_names[1], 'skinModified1')
     env.step(action, 1)
     self.assertEqual(len(env.events()), 0)
     env.reset()
-    self.assertEqual(len(env.events()), 1)
+    self.assertEqual(len(env.events()), 2)
+
 
 if __name__ == '__main__':
   if 'TEST_SRCDIR' in os.environ: